refactor: spit slack sender and slack template
