+++ apiGitHub/index.js
@@ -1,7 +1,8 @@
 const { Octokit } = require("@octokit/rest");
 const { filterPullRequestInformation } = require('./utils');
+const slackPullRequestMessage = require('../slackIntegration/templates/slackPullRequestMessage');
 
-module.exports = (slack, owner, auth, repositories) => {
+module.exports = (slackSender, owner, auth, repositories) => {
     const octokit = new Octokit({
         auth,
         baseUrl: 'https://api.github.com',
@@ -18,19 +19,29 @@ module.exports = (slack, owner, auth, repositories) => {
 
                     if(pullRequests.data.length > 0) {
                         const review = await pullRequests.data
-                        .filter(pullRequest => !pullRequest.draft)
-                        .map(async (pullRequest) => {
-                            const reviewes = await octokit.pulls.listReviews({
-                                owner,
-                                repo,
-                                pull_number: pullRequest.number
-                            });
-                            return filterPullRequestInformation(pullRequest, reviewes.data);
-                        });
-    
+                            .filter(pullRequest => !pullRequest.draft)
+                            .reduce(async (_acc, pullRequest) => {
+                                try {
+                                    const reviewers = await octokit.pulls.listReviews({
+                                        owner,
+                                        repo,
+                                        pull_number: pullRequest.number
+                                    });
+
+                                    const data = filterPullRequestInformation(pullRequest, reviewers.data);
+                                    const acc = await _acc;
+                                    await acc.push(data);
+                                    return acc;
+                                }
+                                catch (err) {
+                                    console.log(err);
+                                }
+                            }, []);
+
                         Promise.all(review)
-                            .then(data => {
-                                slack(data);
+                            .then(async (data) => {
+                                const messages = slackPullRequestMessage(data);
+                                await slackSender(messages);
                             })
                     }
                 };
